+++ code/api/classes/KafkaConsumer/changelog/changelog.md
@@ -1,5 +1,10 @@
 # KafkaConsumer
 
+> **v1.0.2**
+>	* Migliorata la stampa dei topic
+>   * Aggiunto support per stampa topic in un dato intervallo di tempo
+>   * Aggiunto supporto alla Consumazione Multi-Topic dei Record
+
 > **v1.0.1**
 >	* Potenziata documentazione e logica Unit test